+++ src/onegov.core/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Allows to configure the used locales.
+  [msom]
+
 0.40.6 (2017-06-19)
 ~~~~~~~~~~~~~~~~~~~